match debug for D_menu_fmap2_d (#2925)
* menu_fmap2d debug progress

* match debug for d_menu_fmap2D

* fix broken function

* fix weird shield mismatch

* fix other shield mismatch

* fix suggestions

* fix func for wii version

* match func for all versions

* remove cast from saveBitLabels

+++ include/d/actor/d_a_npc_cdn3.h
@@ -223,7 +223,7 @@ public:
                 break;
             }
                 /* dSv_event_flag_c::F_0281 - Shop - Malo Mart opens in Castle Town */
-            if (dComIfGs_isEventBit((u16)dSv_event_flag_c::saveBitLabels[281]) && uVar3 == 9) {
+            if (dComIfGs_isEventBit(dSv_event_flag_c::saveBitLabels[281]) && uVar3 == 9) {
                 uVar3 = 0xb;
             }
         } else {
@@ -257,7 +257,7 @@ public:
                 break;
             }
                 /* dSv_event_flag_c::F_0281 - Shop - Malo Mart opens in Castle Town */
-            if (dComIfGs_isEventBit((u16)dSv_event_flag_c::saveBitLabels[281]) && uVar3 == 10) {
+            if (dComIfGs_isEventBit(dSv_event_flag_c::saveBitLabels[281]) && uVar3 == 10) {
                 uVar3 = 12;
             }
         }
