+++ src/d/actor/d_a_kago.cpp
@@ -522,7 +522,7 @@ void daKago_c::demo_skip(int param_0) {
         setPlayerRideOn();
         field_0x718 = field_0x71a = 0;
         /* dSv_event_flag_c::M_051 - Main Event - Shadow Kargorok (?) (Large) event complete (Horse grass appears in various places) */
-        dComIfGs_onEventBit((u16)dSv_event_flag_c::saveBitLabels[84]);
+        dComIfGs_onEventBit(dSv_event_flag_c::saveBitLabels[84]);
         field_0x748 = 7;
         field_0x74c = NULL;
         break;
